requirements.txt
# ================================================================================
#  Quantum Trade Oracle v2 — Requirements
# ================================================================================

# ── Données & Marché ──────────────────────────────────────────────────────────
yfinance==0.2.35            # Yahoo Finance (actions + crypto)
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
beautifulsoup4==4.12.3      # Scraping RSS/HTML
lxml==5.1.0                 # Parser XML pour les flux RSS
python-dotenv==1.0.0        # Gestion des clés API via .env

# ── Machine Learning Classique ────────────────────────────────────────────────
scikit-learn==1.3.2         # RandomForest, GradientBoosting, scaler
xgboost==2.0.3              # Alternative à GradientBoosting (optionnel)

# ── Deep Learning (PyTorch) ───────────────────────────────────────────────────
# Option 1 — CPU seulement (plus léger)
torch==2.1.2                # LSTM + Transformer
# Option 2 — GPU (si disponible) : pip install torch --index-url https://download.pytorch.org/whl/cu121

# ── NLP & Sentiment ───────────────────────────────────────────────────────────
transformers==4.36.2        # FinBERT (HuggingFace)
vaderSentiment==3.3.2       # Lexique VADER (fallback rapide)
# tokenizers==0.15.0        # Dépendance de transformers

# ── API REST ──────────────────────────────────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.25.0

# ── Dashboard ────────────────────────────────────────────────────────────────
streamlit==1.29.0
plotly==5.18.0

# ── Logging & Utilitaires ────────────────────────────────────────────────────
loguru==0.7.2               # Alternative au logger standard
rich==13.7.0                # Tables et mise en forme console

# ── Tests ────────────────────────────────────────────────────────────────────
pytest==7.4.4
pytest-cov==4.1.0

# ── Optionnel (Reddit via PRAW) ───────────────────────────────────────────────
# praw==7.7.1               # API Reddit officielle
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.36
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pymysql>=1.1.0